Display state, errors don't require immediate attention, and misc. bug fixes
This version allows hte user ot continue other tests while another
tester has an error, and the user can handle or ignore errors
(and handle them later). It also shows the state each tester is in.
Finally, this version includes miscellaneous bug fixes, such as the
program crashing if the user starts a test with a battery ID with
less than six characters.